Backdrop compponent rewritten to TS
